requirements.txt
langchain==0.1.4
langchain_openai==0.0.5
python-dotenv==1.0.1

# for chatbot_oa.py
openai==1.10.0

# for chatbot_hf.py
huggingface_hub==0.20.3
transformers==4.37.1
text-generation==0.6.1
google-search-results==2.4.2
numexpr==2.9.0 
langchainhub==0.1.14
sentencepiece==0.1.99
jinja2==3.1.3